+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii Message Translator Change Log
 
-## 1.0.3 under development
+## 1.1.0 September 07, 2022
 
 - Chg #25: Move `SimpleMessageFormatter` from `yiisoft/translator-formatter-simple` package (@DAGpro, @vjik)